# Multi-threaded media export