--- a/openage/convert/main.py
+++ b/openage/convert/main.py
@@ -1,4 +1,4 @@
-# Copyright 2015-2023 the openage authors. See copying.md for legal info.
+# Copyright 2015-2024 the openage authors. See copying.md for legal info.
 #
 # pylint: disable=too-many-branches
 """
@@ -64,7 +64,7 @@ def convert_assets(
     # add a dir for debug info
     debug_log_path = converted_path / "debug" / datetime.now().strftime("%Y-%m-%d-%H-%M-%S")
     debugdir = DirectoryCreator(debug_log_path).root
-    args.debugdir = AccessSynchronizer(debugdir).root
+    args.debugdir = debugdir
 
     # Create CLI args info
     debug_cli_args(args.debugdir, args.debug_info, args)
@@ -93,9 +93,8 @@ def convert_assets(
     if not data_dir:
         return None
 
-    # make srcdir and targetdir safe for threaded conversion
     args.srcdir = AccessSynchronizer(data_dir).root
-    args.targetdir = AccessSynchronizer(targetdir).root
+    args.targetdir = targetdir
 
     # Create mountpoint info
     debug_mounts(args.debugdir, args.debug_info, args)