… openage/convert/tool/driver.py
… openage/convert/value_object/read/media/drs.py
… openage/util/fslike/filecollection.py